requirements.txt
boto3==1.28.38
botocore==1.31.38
flask==2.3.3
python-dotenv==1.0.0